fix: add delete products from cart method in cart client in order service
feat: add userClient class to get user email from profile service in cart service

+++ EShoppingZone/Profile-Service/src/main/java/com/eshopingzone/profileservice/model/UserProfile.java
@@ -42,7 +42,7 @@ public class UserProfile {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int profileId;
+	private Long profileId;
 
 	@NotBlank
 	@Size(min = 5 , max = 25)
